requirements.txt
asgiref==3.5.2
# backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.3
django-environ==0.9.0
django-heroku==0.3.1
djangorestframework==3.14.0
environ==1.0
gunicorn==20.1.0
idna==3.4
psycopg2==2.9.5
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
stripe==5.0.0
urllib3==1.26.12
whitenoise==6.2.0
